translate: move TranslationSet test helpers to set_test.go
Now that we have a dedicated set_test.go, it makes more sense to put them
there.
